Refactor the sending transactions dialog
