+++ client/app/src/services/transaction-sender.service.js
@@ -2,16 +2,195 @@
   'use strict'
 
   angular.module('arkclient.services')
-    .service('transactionSenderService', ['$mdDialog', 'utilityService', 'accountService', 'transactionBuilderService', TransactionSenderService])
+    .service('transactionSenderService', ['$mdDialog', 'utilityService', 'accountService', 'storageService', 'toastService', 'transactionBuilderService', TransactionSenderService])
 
   /**
    * TransactionSenderService
    * @constructor
+   *
+   * This service is used to send transactions; from displaying the dialog to
+   * validating and executing them.
    */
-  function TransactionSenderService ($mdDialog, utilityService, accountService, transactionBuilderService) {
+  function TransactionSenderService ($mdDialog, utilityService, accountService, storageService, toastService, transactionBuilderService) {
 
     function cancel () {
       $mdDialog.hide()
+    }
+
+    /**
+     * Show the send transaction dialog. Reuses the controller and its $scope TODO
+     */
+    const openDialogIn = ($scope, accountCtrl, selectedAccount) => {
+      const passphrases = accountService.getPassphrases(selectedAccount.address)
+
+      const selectFile = () => {
+        require('electron').remote.dialog.showOpenDialog(fileNames => {
+          if (fileNames === undefined) return
+
+          $scope.send.data.file = fileNames[0]
+        })
+      }
+
+      const parse = filePath => {
+        const fs = require('fs')
+        fs.readFile(filePath, 'utf8', (err, data) => {
+          if (err) {
+            toastService.error('Unable to load file' + ': ' + err)
+          } else {
+            const parse = require('csv-parse')
+            parse(data, { quote: null }, (err, transactions) => {
+              if (err) {
+                return toastService.error('Error while parsing the file')
+              }
+
+              this.showValidateTransactions(selectedAccount, transactions)
+            })
+          }
+        })
+      }
+
+      const submitTransaction = (selectedAccount, formData) => {
+        return transactionBuilderService.createSendTransaction({
+          ledger: selectedAccount.ledger,
+          publicKey: selectedAccount.publicKey,
+          fromAddress: formData.fromAddress,
+          toAddress: formData.toAddress,
+          amount: parseInt(utilityService.arkToArktoshi(formData.amount, 0)),
+          smartbridge: formData.smartbridge,
+          masterpassphrase: formData.passphrase,
+          secondpassphrase: formData.secondpassphrase
+        })
+          .then(transaction => {
+            accountCtrl.showValidateTransaction(selectedAccount, transaction)
+          },
+          accountCtrl.formatAndToastError
+        )
+      }
+
+      function getTotalBalance (fee) {
+        const balance = selectedAccount.balance
+        return utilityService.arktoshiToArk(fee ? balance - fee : balance)
+      }
+
+      function fillSendableBalance () {
+        function setBalance (fee) {
+          const sendableBalance = getTotalBalance(fee)
+          $scope.send.data.amount = sendableBalance > 0 ? sendableBalance : 0
+        }
+        // set the balance immediately, so the user sees something
+        setBalance(accountService.defaultFees.send)
+        // now get the real fees and set it again if necessary
+        accountService.getFees(true).then((fees) => {
+          if (fees.send !== accountService.defaultFees.send) {
+            setBalance(fees.send)
+          }
+        })
+      }
+
+      const submit = tab => {
+        if (!$scope[`${tab}Form`].$valid) {
+          return
+        }
+
+        // In case of data selected from contacts
+        if ($scope.send.data.toAddress.address) {
+          $scope.send.data.toAddress = $scope.send.data.toAddress.address
+        }
+
+        // Remove bad characters before and after in case of bad copy/paste
+        $scope.send.data.toAddress = $scope.send.data.toAddress.trim()
+        $scope.send.data.passphrase = $scope.send.data.passphrase.trim()
+
+        if ($scope.send.data.secondpassphrase) {
+          $scope.send.data.secondpassphrase = $scope.send.data.secondpassphrase.trim()
+        }
+
+        $mdDialog.hide()
+
+        if (tab === 'unique') {
+          submitTransaction(selectedAccount, $scope.send.data)
+        } else if (tab === 'multiple') {
+          submitTransactions(selectedAccount, $scope.send.data)
+        } else {
+          throw new Error(`Unknown tab "${tab}"`)
+        }
+      }
+
+      const querySearch = (text) => {
+        text = text.toLowerCase()
+
+        let accounts = accountCtrl.getAllAccounts()
+        let contacts = storageService.get('contacts') || []
+
+        contacts = contacts.concat(accounts).sort(function (a, b) {
+          if (a.name && b.name) return a.name < b.name
+          else if (a.username && b.username) return a.username < b.username
+          else if (a.username && b.name) return a.username < b.name
+          else if (a.name && b.username) return a.name < b.username
+        })
+
+        return contacts.filter(function (account) {
+          return (account.address.toLowerCase().indexOf(text) > -1) || (account.name && (account.name.toLowerCase().indexOf(text) > -1))
+        })
+      }
+
+      function checkContacts (input) {
+        if (input[0] !== '@') return true
+      }
+
+      const data = {
+        ledger: selectedAccount.ledger,
+        fromAddress: selectedAccount ? selectedAccount.address : '',
+        secondSignature: selectedAccount ? selectedAccount.secondSignature : '',
+        passphrase: passphrases[0] ? passphrases[0] : '',
+        secondpassphrase: passphrases[1] ? passphrases[1] : ''
+      }
+
+      $scope.ac = accountCtrl
+
+      $scope.send = {
+        tab: 'unique',
+        selectFile,
+        data,
+        cancel,
+        submit,
+        checkContacts,
+        querySearch,
+        fillSendableBalance,
+        totalBalance: getTotalBalance(0),
+        remainingBalance: getTotalBalance(0) // <-- initial value, this will change by directive
+      }
+
+      $scope.send.searchTextChange = text => {
+        $scope.send.data.toAddress = text
+      }
+
+      $scope.send.selectedContactChange = contact => {
+        if (contact) {
+          $scope.send.data.toAddress = contact.address
+        }
+      }
+
+      $scope.send.selectTab = tab => {
+        $scope.send.tab = tab
+      }
+
+      $scope.onQrCodeForToAddressScanned = (address) => {
+        // this will trigger the selectedContactChange function, which will then set the toAddress
+        $scope.send.data.selectedAddress = {address: address}
+      }
+
+      $mdDialog.show({
+        parent: angular.element(document.getElementById('app')),
+        templateUrl: './src/components/account/templates/send-transaction-dialog.html',
+        clickOutsideToClose: false,
+        preserveScope: true,
+        scope: $scope
+      })
+    }
+
+    this.showValidateTransactions = function(selectedAccount, transactions) {
+      transactionSenderService.showValidateTransactions($scope, selectedAccount, transactions)
     }
 
     function processTransactionsData(data) {
@@ -88,7 +267,7 @@
     }
 
     return {
-      showValidateTransactions
+      openDialogIn,
     }
   }
